+++ tests/integration/mutation/update/field_kinds/one_to_one/with_simple_test.go
@@ -8,23 +8,196 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package update
+package one_to_one
 
 import (
 	"fmt"
 	"testing"
 
 	testUtils "github.com/sourcenetwork/defradb/tests/integration"
-	fixture "github.com/sourcenetwork/defradb/tests/integration/mutation/one_to_one"
 )
 
-func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromPrimarySide(t *testing.T) {
+// Note: This test should probably not pass, as it contains a
+// reference to a document that doesnt exist.
+func TestMutationUpdateOneToOneNoChild(t *testing.T) {
+	test := testUtils.TestCase{
+		Description: "One to one create mutation, from the wrong side",
+		Actions: []any{
+			testUtils.CreateDoc{
+				CollectionID: 1,
+				Doc: `{
+					"name": "John"
+				}`,
+			},
+			testUtils.Request{
+				Request: `mutation {
+							update_Author(data: "{\"name\": \"John Grisham\",\"published_id\": \"bae-fd541c25-229e-5280-b44b-e5c2af3e374d\"}") {
+								name
+							}
+						}`,
+				Results: []map[string]any{
+					{
+						"name": "John Grisham",
+					},
+				},
+			},
+		},
+	}
+	executeTestCase(t, test)
+}
+
+func TestMutationUpdateOneToOne(t *testing.T) {
+	test := testUtils.TestCase{
+		Description: "One to one update mutation",
+		Actions: []any{
+			testUtils.CreateDoc{
+				CollectionID: 0,
+				Doc: `{
+					"name": "Painted House"
+				}`,
+			},
+			testUtils.CreateDoc{
+				CollectionID: 1,
+				Doc: `{
+					"name": "John Grisham"
+				}`,
+			},
+			testUtils.Request{
+				Request: `
+				mutation {
+					update_Author(data: "{\"name\": \"John Grisham\",\"published_id\": \"bae-3d236f89-6a31-5add-a36a-27971a2eac76\"}") {
+						name
+					}
+				}`,
+				Results: []map[string]any{
+					{
+						"name": "John Grisham",
+					},
+				},
+			},
+			testUtils.Request{
+				Request: `
+					query {
+						Book {
+							name
+							author {
+								name
+							}
+						}
+					}`,
+				Results: []map[string]any{
+					{
+						"name": "Painted House",
+						"author": map[string]any{
+							"name": "John Grisham",
+						},
+					},
+				},
+			},
+			testUtils.Request{
+				Request: `
+					query {
+						Author {
+							name
+							published {
+								name
+							}
+						}
+					}`,
+				Results: []map[string]any{
+					{
+						"name": "John Grisham",
+						"published": map[string]any{
+							"name": "Painted House",
+						},
+					},
+				},
+			},
+		},
+	}
+
+	executeTestCase(t, test)
+}
+
+func TestMutationUpdateOneToOneSecondarySide(t *testing.T) {
+	test := testUtils.TestCase{
+		Description: "One to one create mutation, from the secondary side",
+		Actions: []any{
+			testUtils.CreateDoc{
+				CollectionID: 0,
+				Doc: `{
+					"name": "Painted House"
+				}`,
+			},
+			testUtils.CreateDoc{
+				CollectionID: 1,
+				Doc: `{
+					"name": "John Grisham"
+				}`,
+			},
+			testUtils.Request{
+				Request: `
+				mutation {
+					update_Book(data: "{\"name\": \"Painted House\",\"author_id\": \"bae-2edb7fdd-cad7-5ad4-9c7d-6920245a96ed\"}") {
+						name
+					}
+				}`,
+				Results: []map[string]any{
+					{
+						"name": "Painted House",
+					},
+				},
+			},
+			testUtils.Request{
+				Request: `
+					query {
+						Book {
+							name
+							author {
+								name
+							}
+						}
+					}`,
+				Results: []map[string]any{
+					{
+						"name": "Painted House",
+						"author": map[string]any{
+							"name": "John Grisham",
+						},
+					},
+				},
+			},
+			testUtils.Request{
+				Request: `
+					query {
+						Author {
+							name
+							published {
+								name
+							}
+						}
+					}`,
+				Results: []map[string]any{
+					{
+						"name": "John Grisham",
+						"published": map[string]any{
+							"name": "Painted House",
+						},
+					},
+				},
+			},
+		},
+	}
+	executeTestCase(t, test)
+}
+
+func TestMutationUpdateOneToOne_RelationIDToLinkFromPrimarySide(t *testing.T) {
 	author1Key := "bae-2edb7fdd-cad7-5ad4-9c7d-6920245a96ed"
 	author2Key := "bae-35953caf-4898-518d-9e6b-9ce6cd86ebe5"
 	bookKey := "bae-22e0a1c2-d12b-5bfd-b039-0cf72f963991"
 
 	test := testUtils.TestCase{
-		Description: "One to one update mutation using alias relation id from single side",
+		Description: "One to one update mutation using relation id from single side (wrong)",
 		Actions: []any{
 			testUtils.Request{
 				Request: `mutation {
@@ -53,7 +226,7 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromPrimarySide(t *testin
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
- 						create_Book(data: "{\"name\": \"Painted House\",\"author\": \"%s\"}") {
+ 						create_Book(data: "{\"name\": \"Painted House\",\"author_id\": \"%s\"}") {
  							_key
  							name
  						}
@@ -70,7 +243,7 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromPrimarySide(t *testin
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
- 						update_Author(id: "%s", data: "{\"published\": \"%s\"}") {
+ 						update_Author(id: "%s", data: "{\"published_id\": \"%s\"}") {
  							name
  						}
  					}`,
@@ -82,16 +255,16 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromPrimarySide(t *testin
 		},
 	}
 
-	fixture.ExecuteTestCase(t, test)
+	executeTestCase(t, test)
 }
 
-func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromSecondarySide(t *testing.T) {
+func TestMutationUpdateOneToOne_RelationIDToLinkFromSecondarySide(t *testing.T) {
 	author1Key := "bae-2edb7fdd-cad7-5ad4-9c7d-6920245a96ed"
 	author2Key := "bae-35953caf-4898-518d-9e6b-9ce6cd86ebe5"
 	bookKey := "bae-22e0a1c2-d12b-5bfd-b039-0cf72f963991"
 
 	test := testUtils.TestCase{
-		Description: "One to one update mutation using alias relation id from secondary side",
+		Description: "One to one update mutation using relation id from secondary side",
 		Actions: []any{
 			testUtils.Request{
 				Request: `mutation {
@@ -120,7 +293,7 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromSecondarySide(t *test
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
- 						create_Book(data: "{\"name\": \"Painted House\",\"author\": \"%s\"}") {
+ 						create_Book(data: "{\"name\": \"Painted House\",\"author_id\": \"%s\"}") {
  							_key
  							name
  						}
@@ -137,7 +310,7 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromSecondarySide(t *test
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
- 						update_Book(id: "%s", data: "{\"author\": \"%s\"}") {
+ 						update_Book(id: "%s", data: "{\"author_id\": \"%s\"}") {
  							name
  						}
  					}`,
@@ -149,17 +322,17 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromSecondarySide(t *test
 		},
 	}
 
-	fixture.ExecuteTestCase(t, test)
+	executeTestCase(t, test)
 }
 
-func TestMutationUpdateOneToOne_AliasWithInvalidLengthRelationIDToLink_Error(t *testing.T) {
+func TestMutationUpdateOneToOne_InvalidLengthRelationIDToLink_Error(t *testing.T) {
 	author1Key := "bae-2edb7fdd-cad7-5ad4-9c7d-6920245a96ed"
 	invalidLenSubKey := "35953ca-518d-9e6b-9ce6cd00eff5"
 	invalidAuthorKey := "bae-" + invalidLenSubKey
 	bookKey := "bae-22e0a1c2-d12b-5bfd-b039-0cf72f963991"
 
 	test := testUtils.TestCase{
-		Description: "One to one update mutation using invalid alias relation id",
+		Description: "One to one update mutation using invalid relation id",
 		Actions: []any{
 			testUtils.Request{
 				Request: `mutation {
@@ -176,7 +349,7 @@ func TestMutationUpdateOneToOne_AliasWithInvalidLengthRelationIDToLink_Error(t *
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
- 						create_Book(data: "{\"name\": \"Painted House\",\"author\": \"%s\"}") {
+ 						create_Book(data: "{\"name\": \"Painted House\",\"author_id\": \"%s\"}") {
  							_key
  							name
  						}
@@ -193,7 +366,7 @@ func TestMutationUpdateOneToOne_AliasWithInvalidLengthRelationIDToLink_Error(t *
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
-						update_Book(id: "%s", data: "{\"author\": \"%s\"}") {
+						update_Book(id: "%s", data: "{\"author_id\": \"%s\"}") {
 							name
 						}
 					}`,
@@ -205,16 +378,16 @@ func TestMutationUpdateOneToOne_AliasWithInvalidLengthRelationIDToLink_Error(t *
 		},
 	}
 
-	fixture.ExecuteTestCase(t, test)
+	executeTestCase(t, test)
 }
 
-func TestMutationUpdateOneToOne_InvalidAliasRelationNameToLinkFromSecondarySide_Error(t *testing.T) {
+func TestMutationUpdateOneToOne_InvalidRelationIDToLinkFromSecondarySide_Error(t *testing.T) {
 	author1Key := "bae-2edb7fdd-cad7-5ad4-9c7d-6920245a96ed"
 	invalidAuthorKey := "bae-2edb7fdd-cad7-5ad4-9c7d-6920245a96ee"
 	bookKey := "bae-22e0a1c2-d12b-5bfd-b039-0cf72f963991"
 
 	test := testUtils.TestCase{
-		Description: "One to one update mutation using alias relation id from secondary side",
+		Description: "One to one update mutation using relation id from secondary side",
 		Actions: []any{
 			testUtils.Request{
 				Request: `mutation {
@@ -231,7 +404,7 @@ func TestMutationUpdateOneToOne_InvalidAliasRelationNameToLinkFromSecondarySide_
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
- 						create_Book(data: "{\"name\": \"Painted House\",\"author\": \"%s\"}") {
+ 						create_Book(data: "{\"name\": \"Painted House\",\"author_id\": \"%s\"}") {
  							_key
  							name
  						}
@@ -248,7 +421,7 @@ func TestMutationUpdateOneToOne_InvalidAliasRelationNameToLinkFromSecondarySide_
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
-						update_Book(id: "%s", data: "{\"author\": \"%s\"}") {
+						update_Book(id: "%s", data: "{\"author_id\": \"%s\"}") {
 							name
 						}
 					}`,
@@ -260,16 +433,16 @@ func TestMutationUpdateOneToOne_InvalidAliasRelationNameToLinkFromSecondarySide_
 		},
 	}
 
-	fixture.ExecuteTestCase(t, test)
+	executeTestCase(t, test)
 }
 
-func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromSecondarySideWithWrongField_Error(t *testing.T) {
+func TestMutationUpdateOneToOne_RelationIDToLinkFromSecondarySideWithWrongField_Error(t *testing.T) {
 	author1Key := "bae-2edb7fdd-cad7-5ad4-9c7d-6920245a96ed"
 	author2Key := "bae-35953caf-4898-518d-9e6b-9ce6cd86ebe5"
 	bookKey := "bae-22e0a1c2-d12b-5bfd-b039-0cf72f963991"
 
 	test := testUtils.TestCase{
-		Description: "One to one update mutation using relation alias name from secondary side, with a wrong field.",
+		Description: "One to one update mutation using relation id from secondary side, with a wrong field.",
 		Actions: []any{
 			testUtils.Request{
 				Request: `mutation {
@@ -298,7 +471,7 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromSecondarySideWithWron
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
- 						create_Book(data: "{\"name\": \"Painted House\",\"author\": \"%s\"}") {
+ 						create_Book(data: "{\"name\": \"Painted House\",\"author_id\": \"%s\"}") {
  							_key
  							name
  						}
@@ -315,7 +488,7 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromSecondarySideWithWron
 			testUtils.Request{
 				Request: fmt.Sprintf(
 					`mutation {
- 						update_Book(id: "%s", data: "{\"notName\": \"Unpainted Condo\",\"author\": \"%s\"}") {
+ 						update_Book(id: "%s", data: "{\"notName\": \"Unpainted Condo\",\"author_id\": \"%s\"}") {
  							name
  						}
  					}`,
@@ -327,5 +500,5 @@ func TestMutationUpdateOneToOne_AliasRelationNameToLinkFromSecondarySideWithWron
 		},
 	}
 
-	fixture.ExecuteTestCase(t, test)
+	executeTestCase(t, test)
 }
